refactor: Use IntersectionObserver in Layout component (#168)
This might need a bit of tweaking so consider it a WIP. 

The idea is that instead of calculating the hero offset in each animation frame - we can use the IntersectionObserver API. We are notified each time the intersection changes and when a `.side-nav__title` intersects with the hero banner, we apply some white color. Should offer some performance benefits and might also help solve #86.

Seems like the [browser support is pretty good](https://caniuse.com/#feat=intersectionobserver), so this should work out of the box in most browsers. What are your thoughts on this?