doc: alternative title for landing page (#343)
I'm not a huge fan of the current text:

"The power of JavaScript minus the browser"

This is inaccurate as Node.js is offering a ton of interfaces and
support e.g. system IO that the Browser doesn't offer. Also, while it
is not a browser I do like to think of Node.js as an active participate
in the Web Platform... the comparison here has a negative charge that I
don't think does us any benefit.