Update index.md
Change "(If one of these functions is never called in the execution path, the promise will remain in a pending state)" to "(If none of these functions is called in the execution path, the promise will remain in a pending state)". The first implies that one of them not being called is enough to remain in a pending state, but actually you only need one to be called to exit that state. Examples given below contradict this
````
new Promise((resolve, reject) => {
  reject('Error')
}).catch(err => {
  console.error(err)
})
````