Change Chalk to use v4, as v5 is now Pure ESM
To use `const chalk = require("chalk");` you have to install version 4 of the `chalk` package. Their newest version is pure ESM and will give an error: `(node:61082) Warning: To load an ES module, set "type": "module" in the package.json or use the .mjs extension.`. You could update the documentation to explain how to use ESM, but I don't think this is the right place for that, so I think this is the right change.